+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This page reports the main releases only and the main changes therein.
 
 ### Fixed
 - fixed threshold parameter behavior un unsharp mask filter
+- fixed frames import in retouch mode
 - prevent failure in case a single file is selected for parallel alignment
 - reverts to sequential pyramid focus stacking if temporary disk space is insufficient
 - added protections against invalid file content